fix: update db schema files and ts config. Update README.
