+++ drizzle.config.ts
@@ -2,7 +2,7 @@
 import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
-  schema: "./src/db/schemas/*.ts",
+  schema: "./src/db/schema/schema.ts",
   out: "./src/db/migrations",
   driver: "turso",
   dbCredentials: {
