+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "paisabank-server",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "main": "dist/index.js",
   "type": "module",
   "scripts": {
