+++ src/controllers/auth.ts
@@ -6,7 +6,7 @@ import { eq } from "drizzle-orm";
 // Internal deps
 import { db } from "../db/db.js";
 import { generateToken } from "../lib/token.js";
-import { UsersTable } from "../db/schemas/users.js";
+import { UsersTable } from "../db/schema/schema.js";
 
 export const createUser = async (req: Request, res: Response) => {
   try {
