+++ tsconfig.json
@@ -2,10 +2,10 @@
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
     "noImplicitAny": true,
-    "module": "ES2015",
-    "target": "es5",
+    "module": "NodeNext",
+    "target": "ES2020",
     "rootDir": "./",
-    "moduleResolution": "Bundler",
+    "moduleResolution": "NodeNext",
     "allowJs": true,
     "removeComments": true,
     "outDir": "./dist",
